Inline MutableCallSite targets with LambdaForm
This commit inlines MutableCallSite targets when the MCS is used
directly via invokedynamic. Inlining for MCS dynamic invoker handles
will be implemented afterward.

The compiler will find the target (in mutableCallSiteEpoch()) directly
from the target field rather than "epoch," which doesn't exist in the
LambdaForm-based implementation.

Invokers.getCallSiteTarget() is recognized and InterpreterEmulator
treats it in much the same way as MutableCallSite.getTarget(), except
that it's necessary to check that the call site is in fact an instance
of MutableCallSite (which is always guaranteed for the latter method).

InterpreterEmulator already knows the MCS reference location when it
creates the TR_J9MutableCallSite, so it sets it right away. This stops
inliner from looking through the trees later in an attempt to find the
MCS again.

The inliner policy's shouldRemoveDifferingTargets() is overridden to
avoid removing the refined targets from invokeBasic() even though the
invokeBasic() call is direct. The method symbol can't be refined in the
call node because the known object information is only reliable after
passing the MCS target guard, i.e. only within the inlined body, once
the call has already been inlined.

To that end, make sure to give the callee's LambdaForm method a known
object prex argument for the target handle. Previously the logic to do
so applied only to archetype specimens, i.e. only to the original OpenJ9
implementation of MethodHandle.

When the target of a MutableCallSite changes, if its previous target has
been inlined anywhere, then all guards protecting the inlined body must
be invalidated. This invalidation is combined with the actual update of
the target handle in jitSetMutableCallSiteTarget(), which the VM will
now call rather than set the target directly. The reason for combining
the update with the invalidation is to do them both as part of a single
critical section, which ensures that all guards are patched correctly.
For more details, see the comments in jitSetMutableCallSiteTarget().

A hidden invalidationCookie field is added to MutableCallSite, which
allows mutable call site guards' runtime assumptions to be keyed in the
same way as they were in the original OpenJ9 JSR292 implementation.