Add check-spelling
This is configured to use a matrix (see the readme.md).

Ideally over time a significant portion of the items in the
`allow/*.txt` and `expect/typos.txt` files will be removed.
They're there in order to give a chance to start using this
action before all of the pending spelling fixes are applied.